requirements.txt
mutagen
python3-pyqt5.qtmultimedia
